src/ObjectDetection/requirements.txt
ultralytics==8.1.0
torch
torchvision
pillow
